Refactor SRD ingestion into modular subpackage
