+++ scripts/ingest_srd35/__init__.py
@@ -1,0 +1,11 @@
+from .constants import DEFAULT_MANIFEST
+from .paths import load_manifest
+from .pipeline import ingest_source
+from .rtf_decoder import decode_rtf_text
+
+__all__ = [
+    "DEFAULT_MANIFEST",
+    "decode_rtf_text",
+    "ingest_source",
+    "load_manifest",
+]
